Clearing folders.
