--- 1688InfoToExcelParser/AppFlow/AppDelegate.swift
@@ -1,63 +1,0 @@
-////
-////  AppDelegate.swift
-////  testJSON
-////
-////  Created by oleh yeroshkin on 10.06.2024.
-////
-//
-//import Cocoa
-//
-//class AppDelegate: NSObject, NSApplicationDelegate {
-//
-//    var window: NSWindow!
-//
-//    func applicationDidFinishLaunching(_ aNotification: Notification) {
-//        window = NSWindow(contentRect: NSMakeRect(0, 0, 400, 300),
-//                          styleMask: [.titled, .closable, .resizable, .miniaturizable],
-//                          backing: .buffered,
-//                          defer: false)
-//        window.center()
-//        window.title = "AppKit App"
-//        window.makeKeyAndOrderFront(nil)
-//        window.contentViewController = ViewController()
-//        setupMenu()
-//    }
-//
-//    func applicationWillTerminate(_ aNotification: Notification) {
-//        // Insert code here to tear down your application
-//    }
-//
-//    func applicationSupportsSecureRestorableState(_ app: NSApplication) -> Bool {
-//        return true
-//    }
-//
-//
-//    func setupMenu() {
-//        // Create the main menu
-//        let mainMenu = NSMenu()
-//        NSApp.mainMenu = mainMenu
-//
-//        // Create the application menu
-//        let appMenuItem = NSMenuItem()
-//        mainMenu.addItem(appMenuItem)
-//
-//        let appMenu = NSMenu()
-//        appMenuItem.submenu = appMenu
-//
-//        let appName = ProcessInfo.processInfo.processName
-//        appMenu.addItem(withTitle: "Quit \(appName)", action: #selector(NSApplication.terminate(_:)), keyEquivalent: "q")
-//
-//        // Create the Edit menu
-//        let editMenuItem = NSMenuItem()
-//        mainMenu.addItem(editMenuItem)
-//
-//        let editMenu = NSMenu(title: "Edit")
-//        editMenuItem.submenu = editMenu
-//
-//        editMenu.addItem(withTitle: "Cut", action: #selector(NSText.cut(_:)), keyEquivalent: "x")
-//        editMenu.addItem(withTitle: "Copy", action: #selector(NSText.copy(_:)), keyEquivalent: "c")
-//        editMenu.addItem(withTitle: "Paste", action: #selector(NSText.paste(_:)), keyEquivalent: "v")
-//        editMenu.addItem(withTitle: "Select All", action: #selector(NSText.selectAll(_:)), keyEquivalent: "a")
-//    }
-//}
-//
