+++ 1688InfoToExcelParser/Models/BagModel.swift
@@ -1,0 +1,27 @@
+//
+//  BagModel.swift
+//  1688InfoToExcelParser
+//
+//  Created by Oleh Yeroshkin on 21.06.2024.
+//
+
+import Foundation
+
+struct BagModel {
+    let title: String?
+    let images: [Data]
+    let article: String
+    let aritcleAndColorSku: String?
+    let productName: String? // dress
+    let price: String
+
+    let itemsByColor: [ItemByColor]
+
+    let weight: Float
+    let compositionChinese: String
+    let bagSize: String
+    let fabric: String
+    let strap: String
+}
+
+
