+++ 1688InfoToExcelParser/Views/BagsPropertiesView.swift
