+++ 1688InfoToExcelParser/Views/GetLinkView.swift
