+++ 1688InfoToExcelParser/Views/SizePropertiesCreationView.swift
